+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # CHANGELOG
 
-## 2026-02-07
+## Project Initialized
 
-- chore: 按照项目治理规范初始化 CHANGELOG / Project initialized according to governance specs
+- chore: Project initialized for distribution.